pyproject.toml
[tool.poetry]
name = "vox-agent-os"
version = "0.1.0"
description = "An open-source, voice-first Agent OS for human-computer interaction"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "pkg" }
]

[tool.poetry.dependencies]
python = "<3.13,>=3.10,"
openai = "^1.0.0"
sounddevice = "^0.4.6"
scipy = "^1.11.0"
typer = "^0.9.0"
rich = "^13.0.0"
pydantic = "^2.6.1"
PyPDF2 = "^3.0.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
numpy = "^1.24.0"
wavio = "^0.0.7"
composio = "*"
composio_openai = "*"
crewai = "^0.108.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^23.0"
isort = "^5.0"
mypy = "^1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
vox = "pkg.interfaces.cli.main:app"
soc2 = "pkg.interfaces.cli.soc2_runner:app"